+++ CHANGELOG.md
@@ -168,3 +168,4 @@
 - 2023-11-12: Add README, NOTICE, follow up, mod, +5 db
 - 2023-11-13: Refactor LICENSE, main, find email, mod, +6
 - 2023-11-13: Valid follow up, campaigns, task, find email, +5
+- 2023-11-14: Features follow up, qualify lead, sender, mod, +4